polish(app): move .spool drop target to SharesPage
Imports land as drafts, so the drop affordance belongs next to the
drafts list, not on the Library home. SharesPage now consumes
useSpoolDrop directly; LibraryLanding is back to a plain session list
with no drop wiring. Drop overlay border softened from 2px solid-dashed
accent to 1px accent/70 dashed + bg/60 — the overlay is a hint, not a
wall.

useSpoolDrop gains an optional onReject(files) so callers can surface
their own UI when a drop arrives without a .spool. (Wiring + toast copy
land in the follow-up toast unification PR.)
