+++ packages/app/src/renderer/components/LibraryLanding.tsx
@@ -1,14 +1,12 @@
-import { useCallback, useEffect, useMemo, useState, type ReactNode } from 'react'
+import { useEffect, useMemo, useState, type ReactNode } from 'react'
 import type { Session } from '@spool-lab/core'
 import SessionRow from './SessionRow.js'
-import { useSpoolDrop } from '../hooks/useSpoolDrop.js'
 
 type Props = {
   onSelectProject: (identityKey: string) => void
   onOpenSession: (uuid: string) => void
   onCopySessionId: (source: Session['source']) => void
   onShare?: (uuid: string) => void
-  onImportSpool?: (file: File) => void | Promise<void>
 }
 
 type DateBucket = {
@@ -16,15 +14,7 @@ type DateBucket = {
   sessions: Session[]
 }
 
-export default function LibraryLanding({ onOpenSession, onCopySessionId, onShare, onImportSpool }: Props) {
-  const onImport = useCallback(
-    (file: File) => onImportSpool?.(file),
-    [onImportSpool],
-  )
-  const { isDragActive, dragHandlers } = useSpoolDrop({
-    enabled: Boolean(onImportSpool),
-    onImport,
-  })
+export default function LibraryLanding({ onOpenSession, onCopySessionId, onShare }: Props) {
   const [pinnedSessions, setPinnedSessions] = useState<Session[]>([])
   const [recentSessions, setRecentSessions] = useState<Session[] | null>(null)
   const [reloadKey, setReloadKey] = useState(0)
@@ -76,12 +66,7 @@ export default function LibraryLanding({ onOpenSession, onCopySessionId, onShare
   const totalSessions = (pinnedSessions.length) + (recentSessions?.length ?? 0)
 
   return (
-    <div
-      data-testid="library-landing"
-      className="relative flex flex-col h-full overflow-hidden"
-      {...dragHandlers}
-    >
-      {isDragActive && <SpoolDropOverlay />}
+    <div data-testid="library-landing" className="flex flex-col h-full overflow-hidden">
       <div className="flex-1 overflow-y-auto pb-12 [mask-image:linear-gradient(to_bottom,black_calc(100%_-_24px),transparent)]">
         {recentSessions === null ? (
           <SessionRowsSkeleton count={6} />
@@ -176,20 +161,6 @@ export function CollapsibleSection({
         </svg>
       </button>
       {open && children}
-    </div>
-  )
-}
-
-function SpoolDropOverlay() {
-  return (
-    <div
-      data-testid="library-spool-drop-overlay"
-      aria-hidden
-      className="absolute inset-2 z-20 pointer-events-none flex items-center justify-center rounded-[10px] border-2 border-dashed border-accent dark:border-accent-dark bg-accent-bg/70 dark:bg-accent-bg-dark/70 backdrop-blur-[1px]"
-    >
-      <p className="text-sm font-medium text-accent dark:text-accent-dark">
-        Drop <span className="font-mono">.spool</span> to import
-      </p>
     </div>
   )
 }
